% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConfigFile.R
\name{checkEmptyValue}
\alias{checkEmptyValue}
\title{checkEmptyValue}
\usage{
checkEmptyValue(parameter)
}
\arguments{
\item{parameter}{to check}
}
\value{
boolean value
}
\description{
This utility function check whether the existing provided parameter is null, n.a. or empty
}
