% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConfigFile.R
\name{checkNamedParameter}
\alias{checkNamedParameter}
\title{checkNamedParameter}
\usage{
checkNamedParameter(paramList)
}
\arguments{
\item{paramList}{list of named parameter}
}
\value{
list corrected with default values if the parameter is missing or empty
}
\description{
This function takes a list of named parameter and check on the compulsory
parameter or optional parameter. In the latter case it sets defaults values
with a dedicated warning
}
\details{
Compulsory parameter (existing or not null):
\itemize{
\item datasetDir: directory with all the raw datasets
\item fieldBookFilename: spreadsheet with all the metadata for the processing
\item id: id of the individual element monitored (plant, leaf) in the fieldbook file
\item date: date of the sample in the fieldbook file
\item start_time: starting time of the sample in the fieldbook file
\item ending_time: ending time of the sample in the fieldbook file
\item spectrum_nb: number of spectrum collected for this id colname in the fieldbook file
}

Optional parameter with default: if missing or empty a default value is set
\itemize{
\item destSubDir: directory to write the input  (./out/)
\item outputFilename: filename output (campaign_sumup.xlsx)
}

Optional parameter: if missing no warning and if empty a warning is set
\itemize{
\item location: location colname in the fieldbook file ("NULL")
\item block: block or replica  colname in the fieldbook file ("NULL")
\item factor1: first factor or treatment  of the experiment in the fieldbook file ("NULL")
\item factor2: second factor or treatment of the experiement in the fieldbook file ("NULL")
}
}
