% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramanFileLoad.R
\name{extractMetadataFromFileName}
\alias{extractMetadataFromFileName}
\title{extractMetadataFromFileName}
\usage{
extractMetadataFromFileName(filename)
}
\arguments{
\item{filename}{the filename}
}
\value{
dataframe containing the metadata
}
\description{
This function transform the filename into the metadata of the file
}
