% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRamanData.R
\name{extractRamanData}
\alias{extractRamanData}
\title{extractRamanData}
\usage{
extractRamanData(fieldBook, datasetDir)
}
\arguments{
\item{fieldBook}{tibble with fieldbook information for file recover}

\item{datasetDir}{directory where the raman datafile are stored}
}
\value{
the dataset file produced
}
\description{
main function to process and extract the data
}
