% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldBookLoad.R
\name{fieldBookLoad}
\alias{fieldBookLoad}
\title{fieldBookLoad}
\usage{
fieldBookLoad(fieldBookAbsolutePath, fieldBookMetadata)
}
\arguments{
\item{fieldBookAbsolutePath}{absolute path of the field book}

\item{fieldBookMetadata}{fieldbook metadata}
}
\value{
tibble containing the field book content (need further post processing)
}
\description{
This function load the field book file and rename the column
}
