% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramanFileLoad.R
\name{intersectFileWithDate}
\alias{intersectFileWithDate}
\title{intersectFileWithDate}
\usage{
intersectFileWithDate(fieldBookTibble, dataFileTibble)
}
\arguments{
\item{fieldBookTibble}{tibble containing the metadata from the fieldbook require column
date, startTime, endTime, spectrumNb, id all other column are reported}

\item{dataFileTibble}{tibble conatining the filename and metadata extracted form it. Requires columns
date,time all other column are reported}
}
\value{
tibble with the merged metadata according to the time
}
\description{
This function intersect the file corresponding to the fieldbook data considering
day, month and time
}
\details{
CAUTION due to manipulation error only the day and month of the filename is considered
}
