% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramanFileLoad.R
\name{loadRamanDataFile}
\alias{loadRamanDataFile}
\title{loadRamanDataFile}
\usage{
loadRamanDataFile(filepath, filename)
}
\arguments{
\item{filepath}{the filepath of the data file}

\item{filename}{the unique identifier for the file}
}
\value{
dataframe with the filename and all wavelength as column
}
\description{
This function loads the data from the raw raman datafile with the following structrure
first column correspond to wavelength and second column the value
All the value are transposed in a dataframe with the column name
corresponding to the first column of the file rounded
The filename is indicated also as an identifier
}
