% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldBookLoad.R
\name{p_secondInOneDay}
\alias{p_secondInOneDay}
\title{p_secondInOneDay : global constant representing second in one day}
\usage{
p_secondInOneDay
}
\description{
p_secondInOneDay : global constant representing second in one day
}
