% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramanFieldDataExtraction.R
\name{ramanFieldDataExtraction}
\alias{ramanFieldDataExtraction}
\title{ramanFieldDataExtraction
Main function to match and extract Raman file based on the fieldbook}
\usage{
ramanFieldDataExtraction(configFilePathname = "./config.yml")
}
\arguments{
\item{configFilePathname}{filepath of the config file (default ./config.yml)}
}
\value{
dataset written
}
\description{
ramanFieldDataExtraction
Main function to match and extract Raman file based on the fieldbook
}
