% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldBookLoad.R
\name{renameCompulsoryParameter}
\alias{renameCompulsoryParameter}
\title{renameCompulsoryParameter}
\usage{
renameCompulsoryParameter(fieldBookdData, fieldBookMetadata)
}
\arguments{
\item{fieldBookdData}{tibble with compulsory parameter's config file name}

\item{fieldBookMetadata}{named list with fieldbook user column name}
}
\value{
tibble with with compulsory parameter's script name
}
\description{
This function takes rename in a tibble the compulsory parameter
}
